+++ menu/menu_animation.c
@@ -466,7 +466,7 @@ bool menu_animation_update(animation_t *animation, float dt)
  * Take the contents of @str and apply a ticker effect to it,
  * and write the results in @buf.
  **/
-void menu_animation_ticker_line(char *buf, size_t len, unsigned idx,
+void menu_animation_ticker_line(char *buf, size_t len, uint64_t idx,
       const char *str, bool selected)
 {
    unsigned ticker_period, phase, phase_left_stop;
